Extract traceback code.
Add a missing aria-label.
